requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.3.0
Flask-WTF==1.2.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.8
pyodbc==5.1.0
gunicorn==20.1.0
python-dotenv==1.0.1
requests==2.28.2
Jinja2==3.1.2
itsdangerous==2.2.0
click==8.1.7
Werkzeug>=3.0.0
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0-py3-none-any.whl#sha256=a3a8eed42c600af17c70ed3c8720ffe2879b5353055e66fe0c998306329052fd
spacy>=3.5,<4.0
blis==0.7.11
catalogue==2.0.10
cymem==2.0.8
preshed==3.0.9
thinc==8.2.5
wasabi==0.10.1
numpy==1.26.4
packaging==21.3
pandas==1.5.0
pymssql==2.2.5
Flask-Cors
